实现double和long双局部变量

+++ BDLCompiler/src/test/java/com/blocklynukkit/bedrockLang/test/LearnByteCode.java
@@ -3,7 +3,7 @@ package com.blocklynukkit.bedrockLang.test;
 import java.util.Objects;
 
 public final class LearnByteCode {
-    public static boolean test(float x, int i) {
+    public static boolean test(double x, int i) {
         return x == i;
     }
 }
